Added hold to throw up food with increasing speed
