--- Capstone-Game/Assets/Scripts/Player/FoodHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8d49162572d3ed84c98fa66a642f051b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
