+++ Capstone-Game/Assets/Scripts/Player/SquirrelFoodGrabber.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f1d5ecc02192ec640ad1ce0e324ce407
+guid: 8d49162572d3ed84c98fa66a642f051b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
